^R Add shared policy-bundle loader and migrate injection (all suites pass unchanged; part 1 of 2, behavior-preserving hooks)
Introduce internal/injectionpolicy/bundle_load.go: the shared recursive
policy-bundle loader with LoadOptions{Parse, ValidateInclude,
ValidateMerged, RebasePath, Clone, StrictVersion} hooks that carry the
only real per-caller divergences. Migrate internal/injection to it;
internal/authresolve and internal/authpolicy follow in part 2 so each PR
stays reviewer-sized (the combined change trips the PR-shape line limit).

Injection behavior is preserved exactly: UTF-8 guard, include-validation
posture, version handling, and error ordering live in injection's own
hooks; its test suite passes byte-unchanged. The frozen public contract
now reads the accepted table names from the shared loader
(internal/metadatautil/public_contract.go + policy/public-contract.toml
comment line); the allowedRootPolicyKeys gate literal stays pinned in
render_injection_bundle.go.

Evidence: go test ./... 47 packages ok; injection/injectionpolicy/
metadatautil/authresolve/authpolicy suites green with the old copies
still in place.

Pre-commit pin-freshness gate bypassed for this unrelated change via its
sanctioned knob; provider bump dispatched via upstream-refresh workflow
(run 33998440605).
